Handle imports that aren't seqs
The current implementation assumes that each of the items in your
`:require` and `:import` forms are seqable, but that may not be the case,
at which point the poly_cli will throw an exception and quit.

This commit works around that by checking explicitly for whether the
import is seqable, and doing the right thing for each case.