Fix NPE when taking the size of a directory (#323)
size calls itself and sums the results, but size can return nil in
some circumstances. For example, a broken symlink will have a size of
nil. This results in the poly tool crashing at startup.

The fix adds an (or ,,, 0) form around the sub-call of size

Thanks for this fix @john-shaffer!