Review and update Introduction
Moved link to cljdoc to a tip that is only displayed if not already
reading on cljdoc. Cljoc link adjusted to point to article we are
viewing (introduction) instead of readme.

Same content but re-organized.

Starts with goal, then show the details to further entice reader.

Used bullet points to make text easier to digest.

Tried to be more succinct by getting rid of unnecessary words or
rephrasing to use less words.