Review and update Development doc (#341)
Specific instructions are for Cursive only, make that clearer with headings.

Otherwise just edited for clarity.

Notes:
- I notice extra spacing added under images. I removed it. We can
address image/text spacing in cljdoc if we need to.
- I expect the advice Cursive specific instructions on setting up
a dev namespaces are repeated elsewhere? If not, we might want to come
back and generalize this advice for non-Cursive users.
- As of this writing the Cursive docs on tools.deps setup might be
stale, but that's not something under our domain of control.