Review and update Git and Tagging docs
The Git doc seemed to focus on italicized "stable point in time".
It seemed important, so I carried the term through to the Tagging doc.

While walking through Git doc examples, it became obvious to me that I
had missed the tutorial step of creating lisa.clj from the Development
doc. This was because the step was buried in Cursive setup instructions.
I've teased out a Generic Setup section for non Cursive users.

Changed git log cmd in some cases for a less wide and more doc friendly output.

And: a bunch of edits for clarity.