Issue 318f (#371)
* Review and update Tap doc (#367)

Edits for clarity.

Not sure this is in the right spot in the cljdoc tree, but will leave it
where it is for now.

* When creating a workspace, let the :poly alias use the current version of the polylith/clj-poly library, as a Maven dependency.

* Updated production systems.

* Update the game of life image to a version without the *.

* Review and Update Tools.deps doc (#368)

The clojure vs clj discussion now also covers running clojure as a tool.

Now linking to install docs instead of reiterating what was said
therein.

Moved startup time note to install docs.

Also: some end-of-line whitespace trimming (I think I accidentally added
this on my first edits).

* Review and update Build doc (#369)

* Review and update Build doc

Mention tutorial.

Mention what Clojure tools.build is.

Add call-outs to build.clj.

Remove `io.github.seancorfield/build-uber-log4j2-handler` from docs and
example build.clj. It is a useful lib but not specifically needed for
our example.

Point to `build.clj` associated with release instead of master HEAD version.

Don't assume everybody is using Cursive: move help to tip.

Edits for clarity.

* review feedback: does not -> doesn't

* review feedback: update scripts build.clj too

* Review and update Git and Tagging docs (#370)

The Git doc seemed to focus on italicized "stable point in time".
It seemed important, so I carried the term through to the Tagging doc.

While walking through Git doc examples, it became obvious to me that I
had missed the tutorial step of creating lisa.clj from the Development
doc. This was because the step was buried in Cursive setup instructions.
I've teased out a Generic Setup section for non Cursive users.

Changed git log cmd in some cases for a less wide and more doc friendly output.

And: a bunch of edits for clarity.

---------

Co-authored-by: Lee Read <lee@dlread.com>