Review & Update Flags, Testing, Test Runners docs
General
- Noticed I had accidentally changed `command-line` project alias from
`cl` to `cli`, so addressed that in project.adoc.
- Edits for clarity.
- Info snapshot images are now often annotated, usually just with an
orange box to focus the reader on the part of the output we are describing.
Since the snapshots images are generated automatically via a script,
I've decided to link to the generated image from the manually annotated
InkScape svg. I'll circle back in a future PR to look at the script that
generates these snapshots and see what we can improve.

Flags doc
- Describe flags in general
- Add section headings and doc
- Re-order brick flags content to match order of project flags
- Do more hand-holding in flags with the help of new `info` output images
that point to what we are describing.

Testing doc
- Added some upfront description on how poly runs tests which should
help newcomers.
- Annotated info images with focus boxes
- Fixed a fn name typo in setup/teardown examples

Test Runners doc
- re-orged to cover what the average reader might be first interested in
- used Sean's and Imre's test runners in examples so reader can actually
try something concrete out.
- clarified that poly only includes an in-process test runner