Review and update Libraries doc
Generate libs, libs :compact, and libs with overriden dep from
create_example.clj.

Add support to create-example to run a subset of tasks by a group id.
Running all tasks can take several minutes, so it is handy to be able
to, for example, run `bb create-example realworld` to only run the
realword tasks. If no group ids are specified, all tasks are run.

Added more sections and a table of contents.

The word "dependencies" is overloaded because we call a brick a
"dependency" but a 3rd-party library is also a "dependency". I tried to
soften the confusion by adding a few words.

Re-organized content to start with something the reader can easily try
out then dug progressively into details.

Edits for clarity.