Review and update Validations doc
Re-organize and re-section.

Add some info on exit codes.

Add a screenshot to show an example of `ws get:messages`.

Update Warning 206 screenshot to what it should be now: Error 111.

Turf note on Warning 207 that it is only be applied to check command, it's
not true anymore.

Configuration doc:
- the only reason we documented `:bricks` was to describe disabling
Error 111, this is no longer a feature so we no longer need to describe `:bricks`.
- correct  description for `:necessary` and add link to Validations doc.

Edits for clarity.