Review and Update Source Code doc
Added table of contents.

The first block of text talks specifically about multiple paths but had no
heading, so added one.

I downplayed .clj vs .cljc code organization in first section to
be an illustrative of multiple paths. (To me, the text seemed to be
recommending separate `clj` and `cljc` dirs was some sort of poly
recommendation, which felt odd.)

Reworded a some text under the frontend section.
The term "stripped" did not mean much to me, so I tried to instead
explain.
The term "import" was problematic because it will be confused with
Clojure's `import` so reworded that too.

Other edits for clarity.