Review and Update Naming doc (#407)
issue #318

Add a quote on naming.

Add links to other docs.

Naming is subjective. Change tone from "this is a good way to name"
to "consider naming like so".

Reorder advice on multiple interfaces with same name to come after more
generic advice on components.

"For tools, we can use the tool name" did not give any naming advice at
all, so replaced it with a suggestion.

Edits for clarity.