Store configuration in separate config files (#422)
Add support for storing configuration in config.edn files in projects and bricks, issue #315. The 'migrate' command can be used to automatically migrate a workspace. If a workspace is not migrated, then warning 206 is shown. The internal workspace structure is also affected, and configuration is now stored in each entity (project/brick). Profiles are stored under the :profiles key at the root in the workspace structure. For details, check the next-release.txt at the root of the repository for details. The 'create project' command now supports passing in an alias.