Handle reader conditionals correctly
When we read the source files, we are interested in the require and import statements in the namespace declaration. We use these statements to determine which interface and library dependencies each component has. When there are reader conditionals within these statements, and the workspace is configured to use both `clj` and `cljs` dialects, we need to include imports and requires for both `clj` and `cljs`.

We read the source files using edamame. We can pass an option (`:features`) to edamame to choose between reader conditionals when reading the source files. However, when passing both `:clj` and `:cljs` to edamame, it takes the first match from the conditional instead of both. Luckily, it has another option, `:read-cond`, and it accepts a function that receives the complete conditional statement, and we can decide what to return in place of the conditional.

I implemented a function to read both `clj` and `cljs` parts of the reader conditional when both `clj` and `cljs` dialects are enabled for the workspace. Otherwise, I use `:read-cond :allow` to rely on edamame's default implementation.